refactor: Move the progress bar logic into external file for improved optimization
